feat: refactor home page
